Moving Cell embedders to topology
