+++ src/DGtal/topology/CanonicCellEmbedder.ih
