+++ src/DGtal/topology/CanonicDigitalSurfaceEmbedder.ih
