+++ src/DGtal/topology/CanonicSCellEmbedder.ih
