+++ tests/geometry/surfaces/testNormalVectorEstimatorEmbedder.cpp
@@ -34,7 +34,7 @@
 #include <iterator>
 #include "DGtal/base/Common.h"
 #include "ConfigTest.h"
-#include "DGtal/kernel/CanonicDigitalSurfaceEmbedder.h"
+#include "DGtal/topology/CanonicDigitalSurfaceEmbedder.h"
 #include "DGtal/topology/DigitalSurface.h"
 #include "DGtal/topology/DigitalSetBoundary.h"
 #include "DGtal/topology/ImplicitDigitalSurface.h"
