+++ tests/kernel/testEmbedder.cpp
@@ -34,9 +34,9 @@
 #include "DGtal/kernel/CPointEmbedder.h"
 #include "DGtal/kernel/CWithGradientMap.h"
 #include "DGtal/kernel/CanonicEmbedder.h"
-#include "DGtal/kernel/CanonicCellEmbedder.h"
-#include "DGtal/kernel/CanonicSCellEmbedder.h"
-#include "DGtal/kernel/CanonicDigitalSurfaceEmbedder.h"
+#include "DGtal/topology/CanonicCellEmbedder.h"
+#include "DGtal/topology/CanonicSCellEmbedder.h"
+#include "DGtal/topology/CanonicDigitalSurfaceEmbedder.h"
 #include "DGtal/kernel/RegularPointEmbedder.h"
 #include "DGtal/topology/CCellEmbedder.h"
 #include "DGtal/topology/CSCellEmbedder.h"
